Add relative path snapshot builder
We already had a builder for generating snapshots, but it was closely
tied to the snapshot data model.  This patch adds a different builder
for when you have a single list of files with nested relative paths for
each.  The “relative path” builder will automatically construct the
necessary trees as needed as the files are encountered and added.